refactor(validation): flags in ValidateInstruction
